k8s/bootstrap/platform/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "infractl"
version = "0.1.0"
description = "CLI tool for managing stack-ai infrastructure deployments"
requires-python = ">=3.10"
dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "boto3>=1.34.0",
    "pyyaml>=6.0.0",
    "requests>=2.31.0",
    "urllib3>=2.0.0",
    "kubernetes>=29.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "types-requests>=2.31.0",
    "types-PyYAML>=6.0.0",
    "boto3-stubs[sts,ec2,eks,iam]>=1.34.0",
]

[project.scripts]
infractl = "infractl.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["infractl"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
strict = true
